refactor: InputParser 코드 수정 및 패키지 이동
InputParser 중복 코드 수정 및 패키지를 util.parser 밑으로 이동
ErrorMessage도 util 밑으로 이동
